Refactor: Improve server code structure
